requirements.txt
# python version = 3.9.7
absl-py==0.12.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
antlr4-python3-runtime==4.8
anyio==3.4.0
appdirs==1.4.4
argon2-cffi==21.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.1
attrs==21.2.0
audioread==2.1.9
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.6.0
black==21.12b0
bleach==4.1.0
branca==0.4.2
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.8
clang==5.0
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
cmake==3.21.3
colorama==0.4.4
# conda==4.11.0
# conda-build==3.21.7
# conda-package-handling==1.7.3
cryptography==36.0.0
cssselect==1.1.0
customized-konlpy==0.0.64
cycler==0.11.0
Cython==0.29.24
dataclasses==0.6
datasets==1.14.0
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
delayed==0.11.0b1
Deprecated==1.2.13
dill==0.3.4
dlib==19.22.1
docker-pycreds==0.4.0
docutils==0.18.1
editdistance==0.6.0
efficientnet==1.0.0
efficientnet-pytorch==0.7.1
entrypoints==0.3
essential-generators==1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.4.0
# Fiona==1.8.20
Flask==2.0.2
flask-ngrok==0.0.25
flatbuffers==1.12
folium==0.12.1
fonttools==4.28.3
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
fvcore==0.1.5.post20211023
gast==0.4.0
gensim==4.1.2
geopandas==0.10.1
gitdb==4.0.11
GitPython==3.1.43
glob2==0.7
google-api-core==2.2.2
google-api-python-client==2.26.1
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-vision==2.5.0
google-pasta==0.2.0
googleapis-common-protos==1.53.0
googletrans==3.0.0
grpcio==1.42.0
grpcio-status==1.42.0
h11==0.9.0
h2==3.2.0
h3==3.7.3
h5py==3.1.0
hiredis==2.0.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
huggingface-hub==0.0.19
hydra-core==1.1.1
hyperframe==5.2.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.5
imantics==0.1.12
imbalanced-learn==0.8.1
imgaug==0.4.0
implicit==0.4.8
importlib-metadata==4.8.2
importlib-resources==5.4.0
iopath==0.1.9
ipykernel==6.6.0
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
jieba3k==0.35.1
Jinja2==3.0.3
joblib==1.1.0
JPype1==1.3.0
jsonpickle==2.0.0
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-desktop-server==0.1.3
jupyter-server==1.13.1
jupyter-server-proxy==3.2.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Applications==1.0.8
keras-ocr==0.8.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
konlpy==0.5.2
labelme2coco==0.1.2
libarchive-c==3.2
librosa==0.8.1
lightgbm==3.3.0
llvmlite==0.36.0
lmdb==1.2.1
loguru==0.5.3
lxml==4.6.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
matplotlib-venn==0.11.6
mecab-python==1.0.0
mido==1.2.10
missingno==0.5.0
mistune==0.8.4
moviepy==1.0.3
msgpack==1.0.3
multidict==5.2.0
multiprocess==0.70.12.2
munch==2.5.0
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
newspaper3k==0.2.8
nltk==3.6.5
notebook==6.1.6
numba==0.53.1
numpy==1.21.4
oauthlib==3.1.1
omegaconf==2.1.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pip==21.3.1
pixellib==0.7.1
pkginfo==1.8.1
platformdirs==2.4.0
pmdarima==1.8.3
pooch==1.5.2
portalocker==2.3.2
proglog==0.1.9
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.22
proto-plus==1.19.8
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
py4j==0.10.9
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post2
pycosat==0.6.3
pycparser==2.21
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.10.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyproj==3.2.1
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
PySocks==1.7.1
pyspark==3.1.2
pytesseract==0.3.8
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.3
pyvis==0.1.9
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
ray==1.7.0
redis==4.0.2
regex==2021.11.10
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
resampy==0.2.2
rfc3986==1.5.0
rsa==4.8
# ruamel-yaml-conda==0.15.80
sacremoses==0.0.46
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
selenium==4.0.0
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==2.6.0
setproctitle==1.3.3
setuptools==59.4.0
sgmllib3k==1.0.0
Shapely==1.7.1
simpervisor==0.4
six==1.16.0
smart-open==5.2.1
smmap==5.0.1
sniffio==1.2.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.3.1
soynlp==0.0.493
statistics==1.0.3.5
statsmodels==0.13.0
summa==1.2.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-addons==0.14.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.6.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.5.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tinysegmenter==0.3
tldextract==3.1.2
tokenizers==0.10.3
tomli==1.2.2
# torch==1.9.1+cu111
# torchaudio==0.9.1
# torchvision==0.10.1+cu111
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.11.3
trio==0.19.0
trio-websocket==0.9.2
tweepy==3.10.0
typeguard==2.13.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.19
validators==0.18.2
wandb==0.16.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
websockify==0.10.0
Werkzeug==2.0.2
wget==3.2
wheel==0.37.0
widgetsnbextension==3.5.2
wordcloud==1.8.1
wrapt==1.12.1
wsproto==1.0.0
xgboost==1.4.2
xmljson==0.2.1
xxhash==2.0.2
yacs==0.1.8
yarl==1.7.2
zipp==3.6.0
